+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 - Fix accelerator optimized machine type SMT handling.
 - Prefix user visible errors with its source.
+- Fix accelerator optimized machine type socket handling.
 
 ## \[6.1.1\]